Update index.html
Hello and good morning!

This is the latest index.html. I have checked for the website in multiple screen dimensions (14" laptop, 12.4" tablet (horizontal, vertical), 6.67" android smartphone (vertical) and everything looked okay so far. 

The css file linked to this html is "style.css", while I added a "back to top" button via Button.js.

I think the existing images in the images folder can be used as it is. I have edited the html file to find the images in the folder.